requirements.txt
# Lhydra Hybrid Music Recommender System Dependencies

# Core libraries
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep learning
torch>=1.9.0
tqdm>=4.62.0

# Data processing
pyyaml>=6.0
joblib>=1.1.0

# Utilities
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Optional: If using CUDA
# torch-cuda>=11.7 